# CUDA: faster q2_K, q3_K MMQ + int8 tensor cores