This PR overhauls the q2_K and q3_K mul_mat_q kernels and adds int8 tensor core support. Now that all MMQ-supported quantization formats have int8 tensor core support it was possible to simplify the code a little. To make q2_K work I had to implement an alternative data format for q8_1 where instead of the per-block sum the per-halfblock sums are stored (as int8 relative to the max value). The precision loss from doing this seems to be negligible.

`ggml-cuda.cu` now queries and stores the maximum opt-in shared memory per streaming multiprocessor. This is the actual value that should be used as the upper limit for shared memory for kernel launches (needs an additional function call to explicitly raise).

<details>
<summary>Performance vs. master MMQ</summary>

| GPU        | Model             |     Microbatch size | Test     |     t/s master |     t/s cuda-ptx-mma-17 |     Speedup |
| :--------- | :---------------- | ------------------: | :------- | -------------: | ----------------------: | ----------: |
| RTX 4090   | llama 8B Q2_K_M   |                  16 | pp2048   |        1190.56 |                 1702.40 |        1.43 |
| RTX 4090   | llama 8B Q2_K_M   |                  32 | pp2048   |        1588.46 |                 2689.24 |        1.69 |
| RTX 4090   | llama 8B Q2_K_M   |                  64 | pp2048   |        2462.88 |                 4120.64 |        1.67 |
| RTX 4090   | llama 8B Q2_K_M   |                 128 | pp2048   |        2868.68 |                 5635.31 |        1.96 |
| RTX 4090   | llama 8B Q2_K_M   |                 256 | pp2048   |        3553.10 |                 7022.20 |        1.98 |
| RTX 4090   | llama 8B Q2_K_M   |                 512 | pp2048   |        3909.38 |                 7639.35 |        1.95 |
| RTX 4090   | llama 8B Q2_K_M   |                1024 | pp2048   |        4105.15 |                 7755.40 |        1.89 |
| RTX 4090   | llama 8B Q2_K_M   |                2048 | pp2048   |        3936.76 |                 7083.78 |        1.80 |
| RTX 4090   | llama 8B Q3_K_S   |                  16 | pp2048   |         997.96 |                 1611.31 |        1.61 |
| RTX 4090   | llama 8B Q3_K_S   |                  32 | pp2048   |        1346.93 |                 2644.87 |        1.96 |
| RTX 4090   | llama 8B Q3_K_S   |                  64 | pp2048   |        2289.07 |                 4190.48 |        1.83 |
| RTX 4090   | llama 8B Q3_K_S   |                 128 | pp2048   |        2836.75 |                 6014.07 |        2.12 |
| RTX 4090   | llama 8B Q3_K_S   |                 256 | pp2048   |        3710.10 |                 7712.51 |        2.08 |
| RTX 4090   | llama 8B Q3_K_S   |                 512 | pp2048   |        4183.62 |                 8659.58 |        2.07 |
| RTX 4090   | llama 8B Q3_K_S   |                1024 | pp2048   |        4397.62 |                 8644.62 |        1.97 |
| RTX 4090   | llama 8B Q3_K_S   |                2048 | pp2048   |        4248.61 |                 7935.98 |        1.87 |
| RTX 3090   | llama 8B Q2_K_M   |                  16 | pp2048   |         595.56 |                  977.44 |        1.64 |
| RTX 3090   | llama 8B Q2_K_M   |                  32 | pp2048   |         799.47 |                 1378.14 |        1.72 |
| RTX 3090   | llama 8B Q2_K_M   |                  64 | pp2048   |         937.48 |                 1932.08 |        2.06 |
| RTX 3090   | llama 8B Q2_K_M   |                 128 | pp2048   |        1211.96 |                 2414.37 |        1.99 |
| RTX 3090   | llama 8B Q2_K_M   |                 256 | pp2048   |        1516.43 |                 2956.45 |        1.95 |
| RTX 3090   | llama 8B Q2_K_M   |                 512 | pp2048   |        1582.72 |                 3084.91 |        1.95 |
| RTX 3090   | llama 8B Q2_K_M   |                1024 | pp2048   |        1646.13 |                 3196.07 |        1.94 |
| RTX 3090   | llama 8B Q2_K_M   |                2048 | pp2048   |        1628.12 |                 3125.93 |        1.92 |
| RTX 3090   | llama 8B Q3_K_S   |                  16 | pp2048   |         492.46 |                  923.08 |        1.87 |
| RTX 3090   | llama 8B Q3_K_S   |                  32 | pp2048   |         649.50 |                 1315.24 |        2.03 |
| RTX 3090   | llama 8B Q3_K_S   |                  64 | pp2048   |         859.15 |                 1943.93 |        2.26 |
| RTX 3090   | llama 8B Q3_K_S   |                 128 | pp2048   |        1230.25 |                 2646.62 |        2.15 |
| RTX 3090   | llama 8B Q3_K_S   |                 256 | pp2048   |        1543.02 |                 3305.31 |        2.14 |
| RTX 3090   | llama 8B Q3_K_S   |                 512 | pp2048   |        1665.09 |                 3487.42 |        2.09 |
| RTX 3090   | llama 8B Q3_K_S   |                1024 | pp2048   |        1743.56 |                 3575.74 |        2.05 |
| RTX 3090   | llama 8B Q3_K_S   |                2048 | pp2048   |        1735.05 |                 3470.47 |        2.00 |
| RX 6800    | llama 8B Q2_K_M   |                  16 | pp2048   |         140.08 |                  153.15 |        1.09 |
| RX 6800    | llama 8B Q2_K_M   |                  32 | pp2048   |         159.30 |                  197.70 |        1.24 |
| RX 6800    | llama 8B Q2_K_M   |                  64 | pp2048   |         191.10 |                  238.01 |        1.25 |
| RX 6800    | llama 8B Q2_K_M   |                 128 | pp2048   |         228.92 |                  295.19 |        1.29 |
| RX 6800    | llama 8B Q2_K_M   |                 256 | pp2048   |         273.17 |                  349.91 |        1.28 |
| RX 6800    | llama 8B Q2_K_M   |                 512 | pp2048   |         288.55 |                  364.26 |        1.26 |
| RX 6800    | llama 8B Q2_K_M   |                1024 | pp2048   |         277.80 |                  345.47 |        1.24 |
| RX 6800    | llama 8B Q2_K_M   |                2048 | pp2048   |         256.47 |                  311.28 |        1.21 |
| RX 6800    | llama 8B Q3_K_S   |                  16 | pp2048   |         115.97 |                  134.14 |        1.16 |
| RX 6800    | llama 8B Q3_K_S   |                  32 | pp2048   |         127.77 |                  160.74 |        1.26 |
| RX 6800    | llama 8B Q3_K_S   |                  64 | pp2048   |         164.16 |                  206.02 |        1.26 |
| RX 6800    | llama 8B Q3_K_S   |                 128 | pp2048   |         192.03 |                  249.73 |        1.30 |
| RX 6800    | llama 8B Q3_K_S   |                 256 | pp2048   |         228.57 |                  292.16 |        1.28 |
| RX 6800    | llama 8B Q3_K_S   |                 512 | pp2048   |         242.35 |                  306.31 |        1.26 |
| RX 6800    | llama 8B Q3_K_S   |                1024 | pp2048   |         235.30 |                  294.04 |        1.25 |
| RX 6800    | llama 8B Q3_K_S   |                2048 | pp2048   |         220.55 |                  269.35 |        1.22 |
| P40   | llama 8B Q2_K_M |                16 | pp2048 |       273.10 |         314.75 |      1.15 |
| P40   | llama 8B Q2_K_M |                32 | pp2048 |       358.76 |         427.85 |      1.19 |
| P40   | llama 8B Q2_K_M |                64 | pp2048 |       464.29 |         558.64 |      1.20 |
| P40   | llama 8B Q2_K_M |               128 | pp2048 |       534.66 |         677.66 |      1.27 |
| P40   | llama 8B Q2_K_M |               256 | pp2048 |       582.77 |         731.19 |      1.25 |
| P40   | llama 8B Q2_K_M |               512 | pp2048 |       613.71 |         768.46 |      1.25 |
| P40   | llama 8B Q2_K_M |              1024 | pp2048 |       613.27 |         759.92 |      1.24 |
| P40   | llama 8B Q2_K_M |              2048 | pp2048 |       590.71 |         728.91 |      1.23 |
| P40        | llama 8B Q3_K_S   |                  16 | pp2048   |         241.94 |                  249.37 |        1.03 |
| P40        | llama 8B Q3_K_S   |                  32 | pp2048   |         320.23 |                  355.26 |        1.11 |
| P40        | llama 8B Q3_K_S   |                  64 | pp2048   |         452.54 |                  491.01 |        1.09 |
| P40        | llama 8B Q3_K_S   |                 128 | pp2048   |         520.61 |                  572.41 |        1.10 |
| P40        | llama 8B Q3_K_S   |                 256 | pp2048   |         564.99 |                  620.82 |        1.10 |
| P40        | llama 8B Q3_K_S   |                 512 | pp2048   |         592.96 |                  652.16 |        1.10 |
| P40        | llama 8B Q3_K_S   |                1024 | pp2048   |         586.02 |                  643.80 |        1.10 |
| P40        | llama 8B Q3_K_S   |                2048 | pp2048   |         565.22 |                  618.79 |        1.09 |

</details>

<details>
<summary>Performance vs. master cuBLAS</summary>

| GPU        | Model             |     Microbatch size | Test     |     t/s master |     t/s cuda-ptx-mma-17 |     Speedup |
| :--------- | :---------------- | ------------------: | :------- | -------------: | ----------------------: | ----------: |
| RTX 4090   | llama 8B Q2_K_M   |                  16 | pp2048   |        1189.99 |                 1702.40 |        1.43 |
| RTX 4090   | llama 8B Q2_K_M   |                  32 | pp2048   |        1580.39 |                 2689.24 |        1.70 |
| RTX 4090   | llama 8B Q2_K_M   |                  64 | pp2048   |        2454.15 |                 4120.64 |        1.68 |
| RTX 4090   | llama 8B Q2_K_M   |                 128 | pp2048   |        3642.99 |                 5635.31 |        1.55 |
| RTX 4090   | llama 8B Q2_K_M   |                 256 | pp2048   |        5897.75 |                 7022.20 |        1.19 |
| RTX 4090   | llama 8B Q2_K_M   |                 512 | pp2048   |        7797.40 |                 7639.35 |        0.98 |
| RTX 4090   | llama 8B Q2_K_M   |                1024 | pp2048   |        9039.66 |                 7755.40 |        0.86 |
| RTX 4090   | llama 8B Q2_K_M   |                2048 | pp2048   |        8915.95 |                 7083.78 |        0.79 |
| RTX 4090   | llama 8B Q3_K_S   |                  16 | pp2048   |         991.74 |                 1611.31 |        1.62 |
| RTX 4090   | llama 8B Q3_K_S   |                  32 | pp2048   |        1344.66 |                 2644.87 |        1.97 |
| RTX 4090   | llama 8B Q3_K_S   |                  64 | pp2048   |        2287.11 |                 4190.48 |        1.83 |
| RTX 4090   | llama 8B Q3_K_S   |                 128 | pp2048   |        3542.60 |                 6014.07 |        1.70 |
| RTX 4090   | llama 8B Q3_K_S   |                 256 | pp2048   |        5785.44 |                 7712.51 |        1.33 |
| RTX 4090   | llama 8B Q3_K_S   |                 512 | pp2048   |        7704.78 |                 8659.58 |        1.12 |
| RTX 4090   | llama 8B Q3_K_S   |                1024 | pp2048   |        9015.20 |                 8644.62 |        0.96 |
| RTX 4090   | llama 8B Q3_K_S   |                2048 | pp2048   |        8939.09 |                 7935.98 |        0.89 |
| RTX 3090   | llama 8B Q2_K_M   |                  16 | pp2048   |         585.72 |                  977.44 |        1.67 |
| RTX 3090   | llama 8B Q2_K_M   |                  32 | pp2048   |         778.38 |                 1378.14 |        1.77 |
| RTX 3090   | llama 8B Q2_K_M   |                  64 | pp2048   |         913.46 |                 1932.08 |        2.12 |
| RTX 3090   | llama 8B Q2_K_M   |                 128 | pp2048   |        2274.05 |                 2414.37 |        1.06 |
| RTX 3090   | llama 8B Q2_K_M   |                 256 | pp2048   |        3354.70 |                 2956.45 |        0.88 |
| RTX 3090   | llama 8B Q2_K_M   |                 512 | pp2048   |        3984.17 |                 3084.91 |        0.77 |
| RTX 3090   | llama 8B Q2_K_M   |                1024 | pp2048   |        4692.10 |                 3196.07 |        0.68 |
| RTX 3090   | llama 8B Q2_K_M   |                2048 | pp2048   |        4739.62 |                 3125.93 |        0.66 |
| RTX 3090   | llama 8B Q3_K_S   |                  16 | pp2048   |         477.93 |                  923.08 |        1.93 |
| RTX 3090   | llama 8B Q3_K_S   |                  32 | pp2048   |         625.05 |                 1315.24 |        2.10 |
| RTX 3090   | llama 8B Q3_K_S   |                  64 | pp2048   |         829.36 |                 1943.93 |        2.34 |
| RTX 3090   | llama 8B Q3_K_S   |                 128 | pp2048   |        2102.58 |                 2646.62 |        1.26 |
| RTX 3090   | llama 8B Q3_K_S   |                 256 | pp2048   |        3151.56 |                 3305.31 |        1.05 |
| RTX 3090   | llama 8B Q3_K_S   |                 512 | pp2048   |        3847.53 |                 3487.42 |        0.91 |
| RTX 3090   | llama 8B Q3_K_S   |                1024 | pp2048   |        4594.82 |                 3575.74 |        0.78 |
| RTX 3090   | llama 8B Q3_K_S   |                2048 | pp2048   |        4694.42 |                 3470.47 |        0.74 |

</details>

The performance on Ampere and Ada Lovelace seems to still be suboptimal relative to FP16 cuBLAS GEMM.