@slaren 2024-06-14
I am not sure how big of a problem this is, but extending the `test-backend-ops` tests to include q2_k in `base_types` causes it to fail frequently.

```
  MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=16,k=256,bs=[1,1],nr=[1,1]): [MUL_MAT] NMSE = 0.000539679 > 0.000500000 FAIL
  MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=16,k=256,bs=[10,1],nr=[1,1]): [MUL_MAT] NMSE = 0.000567695 > 0.000500000 FAIL
  MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=16,k=256,bs=[10,1],nr=[2,1]): [MUL_MAT] NMSE = 0.000584948 > 0.000500000 FAIL
  MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=16,k=256,bs=[10,10],nr=[1,1]): [MUL_MAT] NMSE = 0.000510545 > 0.000500000 FAIL
  MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=16,k=256,bs=[10,10],nr=[2,1]): [MUL_MAT] NMSE = 0.000504169 > 0.000500000 FAIL
```