@JohannesGaessler 2024-06-14
Sorry, it seems my testing setup for precision was broken due to https://github.com/ggerganov/llama.cpp/pull/7809 ; I had checked the KL divergence using `perplexity` which was the old binary which due to the name change was not getting removed by `make clean`. And since I got the exact same value (within the printed precision) I assumed the precision loss to be negligible. These are the correct results:

| Model          | imatrix | Code          | PPL                 | KL Divergence vs. FP16 | Mean Δp          |
|----------------+---------+---------------+---------------------+------------------------+------------------|
| LLaMA 3 iq2_M  | WT 10m  | master cuBLAS | 8.598441 ± 0.055100 | 0.325965 ± 0.001606    | -6.467 ± 0.046 % |
| LLaMA 3 q2_K_M | WT 10m  | master cuBLAS | 8.646568 ± 0.055594 | 0.332531 ± 0.001572    | -6.507 ± 0.047 % |
| LLaMA 3 q2_K_M | WT 10m  | master MMQ    | 8.646308 ± 0.055604 | 0.332680 ± 0.001573    | -6.503 ± 0.047 % |
| LLaMA 3 q2_K_M | WT 10m  | PR MMQ        | 8.690794 ± 0.055918 | 0.337893 ± 0.001585    | -6.613 ± 0.047 % |
| LLaMA 3 q2_K_S | WT 10m  | master cuBLAS | 9.321797 ± 0.061532 | 0.403376 ± 0.001787    | -7.137 ± 0.049 % |
| LLaMA 3 q2_K_S | WT 10m  | master MMQ    | 9.322530 ± 0.061534 | 0.403561 ± 0.001788    | -7.146 ± 0.049 % |
| LLaMA 3 q2_K_S | WT 10m  | PR MMQ        | 9.467689 ± 0.062546 | 0.418697 ± 0.001824    | -7.488 ± 0.050 % |
| LLaMA 3 iq2_s  | WT 10m  | master cuBLAS | 9.652453 ± 0.063226 | 0.439268 ± 0.001975    | -8.325 ± 0.052 % |

For q2_K_M I think this is fine but for q2_K_S I would like a better solution; it would also be possible to quantize the activations as q8_1 with a block size of 16 in the first place but that will require comparatively more changes. Since right now q2_K_S is I think not worth using over iq2_M and iq2_S anyways the way I would like to proceed is to merge this PR as-is, optimize MMQ for those quant formats that are already implemented, figure out how to best refactor and simplify the code, and then add the unsupported quantization formats during which I will revisit this.