@slaren 2024-06-14
racecheck reports an error with mmq:
```
========= Error: Race reported between Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f70 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:853
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f10 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f30 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f40 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f50 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f80 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3f90 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3fa0 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========     and Write access at void load_tiles_q2_K<(int)64, (int)4, (bool)1>(const char *, int *, __half2 *, int *, const int &, const int &, const int &)+0x3fb0 in /home/diego/code/llama.cpp/ggml-cuda/mmq.cuh:863 [64 hazards]
=========
```
I also noticed that the precision issue happens rarely happens with low batch sizes (mmvq).