@JohannesGaessler 2024-06-14
>racecheck reports an error with mmq:

In this particular case it doesn't matter because multiple threads write the same value multiple times (that's faster than a conditional statement). But I think I can rewrite the code in such a way that this doesn't happen (at the cost of more register usage which at this point in the kernel is probably irrelevant).

>I also noticed that the precision issue happens rarely happens with low batch sizes (mmvq).

That sounds to me like a different issue. What command are you using for testing?