@slaren 2024-06-14
I'm using `test-backend-ops`. This test never fails the NMSE check (batch size=1):
```
test_cases.emplace_back(new test_mul_mat(GGML_TYPE_Q2_K, GGML_TYPE_F32, 16, 1, 256, { 1,  1}, {1, 1}));
```

But this test fails almost always (batch size=32):
```
test_cases.emplace_back(new test_mul_mat(GGML_TYPE_Q2_K, GGML_TYPE_F32, 16, 32, 256, { 1,  1}, {1, 1}));
```