@ggerganov 2024-06-14
> In this particular case it doesn't matter because multiple threads write the same value multiple times (that's faster than a conditional statement).

I believe this still classifies as UB. But regardless, it would be better to not have any data races even if they are benign in practice