@JohannesGaessler 2024-06-14
The idea I had turned out to be slower than a conditional statement so I used that instead.

>But this test fails almost always (batch size=32):

How does that relate to MMVQ? That kernel is only used for batch sizes <= 8.

Unrelated to that, should we maybe set a seed for the precision tests? (To my understanding the data that is used for testing is currently unseeded RNG output.) I think that would make it easier to differentiate between race conditions and other bugs.