@slaren 2024-06-14
> How does that relate to MMVQ? That kernel is only used for batch sizes <= 8.

Sorry, what I meant was the the issue does not happen when mmvq is used, only with mmq. I am assuming that mmvq also uses the new q8_1 formats and is affected that this changes, but maybe I am wrong.