@JohannesGaessler 2024-06-14
>I am assuming that mmvq also uses the new q8_1 formats and is affected that this changes, but maybe I am wrong.

No, MMVQ uses the exact same data layout that it did prior to my changes.