@slaren 2024-06-14
For me q2_K_M it is about 12% slower with bs=512. I think is still a good tradeoff compared to the memory cost of cuBLAS, since I the only reason to use q2_K in the first place is in memory limited situations. Especially since current models that tend to have a very large output tensors, which increases memory usage.

| GPU         | Model           |   Microbatch size | Test   |   t/s master cuBLAS |   t/s cuda-ptx-mma-17 MMQ |   Speedup |
|:------------|:----------------|------------------:|:-------|-------------:|----------------------:|----------:|
| RTX 3090 Ti |  7B Q2_K_M |                16 | pp1024 |       551.50 |                907.57 |      1.65 |
| RTX 3090 Ti |  7B Q2_K_M |                32 | pp1024 |       729.71 |               1363.34 |      1.87 |
| RTX 3090 Ti |  7B Q2_K_M |                64 | pp1024 |       957.33 |               1986.98 |      2.08 |
| RTX 3090 Ti |  7B Q2_K_M |               128 | pp1024 |      2328.11 |               2711.16 |      1.16 |
| RTX 3090 Ti |  7B Q2_K_M |               256 | pp1024 |      3509.39 |               3357.53 |      0.96 |
| RTX 3090 Ti |  7B Q2_K_M |               512 | pp1024 |      4205.41 |               3703.15 |      0.88 |
| RTX 3090 Ti |  7B Q2_K_M |              1024 | pp1024 |      4659.57 |               3650.46 |      0.78 |